feat: implement dummy functions for non-linux
